cleaned up app by seprating, fixed adding numerous artists songs to playlist functio
